+++ engine/test/environment/options.go
@@ -5,7 +5,10 @@ import (
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 
+	fdatastore "github.com/smartcontractkit/chainlink-deployments-framework/datastore"
+	fdeployment "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 	"github.com/smartcontractkit/chainlink-deployments-framework/engine/test/onchain"
+	"github.com/smartcontractkit/chainlink-deployments-framework/offchain"
 )
 
 // Assign the chain container loader constructors to local variables to allow for stubbing in tests.
@@ -148,6 +151,30 @@ func WithSuiContainerN(t *testing.T, n int) LoadOpt {
 func WithLogger(lggr logger.Logger) LoadOpt {
 	return func(cmps *components) error {
 		cmps.Logger = lggr
+		return nil
+	}
+}
+
+// WithDatastore sets a custom datastore for the environment.
+func WithDatastore(ds fdatastore.DataStore) LoadOpt {
+	return func(cmps *components) error {
+		cmps.Datastore = ds
+		return nil
+	}
+}
+
+// WithAddressBook sets a custom address book for the environment.
+func WithAddressBook(ab fdeployment.AddressBook) LoadOpt {
+	return func(cmps *components) error {
+		cmps.AddressBook = ab
+		return nil
+	}
+}
+
+// WithOffchainClient sets a custom offchain client for the environment.
+func WithOffchainClient(oc offchain.Client) LoadOpt {
+	return func(cmps *components) error {
+		cmps.OffchainClient = oc
 		return nil
 	}
 }